# uicache(1): Speed up -a option