The -f option will enforce the old behaviour.

This works by checking /Applications for .app bundles and comparing it
to the apps that are registered in /Applications and making any
necessary charges.

% time uicache -af
uicache-test -af  0.02s user 0.00s system 0% cpu 1:10.68 total
% time uicache -a
uicache-test -a  0.25s user 0.00s system 85% cpu 0.296 total