@CRKatri 2021-08-24
FYI: This PR doesn't add support for secondary system app locations, like `/private/var/preboot/procursus/Applications`.